HUD: Setup habilidades, Propriedades do player e do personagem local.

+++ Assets/Users/Ale/Character.cs
@@ -8,7 +8,7 @@ public class Character : ScriptableObject
     public GameObject characterPrefab;
     public string characterName, frase, Descricao;
     public int characterIndex, damage, ammo;
-    public float HP;
+    public int HP;
     public Sprite characterIcon;
     public Sprite[] ordenedHabillityIcon;
     public string[] ordenedHabillityName, ordenedHabillityDescription;
